Create findJane.sh
This is a bash script that reads a file called list.txt. In this file there are a bunch of file names. The script should look for all files that contain the name jane. After that the script must create a new txt file called oldFiles.txt and insert the names of the files that contain the name jane into it.